emoji: fix merged unicode
for several of the emoji "words", they are a union of unicode points.
Like flags are the two letters together. Female Detective is a number of
connected symbols. This representation needs to be alltogether.

With this update the unicode points are joined, and require spliting on
the "U+". To help with this and be useful as a library there is a
function `emoji.CodepointToUnicode(string) string` to do this now.

Signed-off-by: Vincent Batts <vbatts@hashbangbash.com>