go: bah, golang and nested modules
this aspect of golang modules is a mess. You can not have a top-level
`package main` that has its own go.mod, and then also export a
sub-directory that has its own go.mod for cleaner dependency imports by
outside importers.

So by removing this, you're subject to the top-level go.mod of the
project. Otherwise the top-level main.go needs to be moved to a nested
./cmd/emojisum/ directory, which is a break for any packagers. 😵

Signed-off-by: Vincent Batts <vbatts@hashbangbash.com>